Make fused AdamW and SGD the default
