+++ tt-train/configs/training_configs/training_shakespeare_llama3_gpt2s_size.yaml
@@ -9,7 +9,7 @@ training_config:
   use_clip_grad_norm: false
   model_config: "configs/model_configs/llama3_gpt2s_size.yaml"
 
-optimizer_config: "configs/optimizer_configs/adamw_fused.yaml"
+optimizer_config: "configs/optimizer_configs/adamw.yaml"
 
 device_config:
   enable_ddp: false
