+++ tt-train/configs/training_configs/training_shakespeare_llama70b_pp_fabric_galaxy.yaml
@@ -9,7 +9,7 @@ training_config:
   clip_grad_norm_max_norm: 1.0
   model_config: "configs/model_configs/llama70b.yaml"
 
-optimizer_config: "configs/optimizer_configs/adamw_fused.yaml"
+optimizer_config: "configs/optimizer_configs/adamw.yaml"
 
 device_config:
   enable_tp: true
