+++ tt-train/configs/training_configs/training_shakespeare_nanogpt_memory_eff.yaml
@@ -7,7 +7,7 @@ training_config:
   max_steps: 5000
   model_config: "configs/model_configs/nanogpt.yaml"
 
-optimizer_config: "configs/optimizer_configs/adamw_fused.yaml"
+optimizer_config: "configs/optimizer_configs/adamw.yaml"
 
 device_config:
   enable_ddp: false
