+++ tt-train/configs/training_configs/training_shakespeare_nanollama3.yaml
@@ -9,7 +9,7 @@ training_config:
   clip_grad_norm_max_norm: 1.0
   model_config: "configs/model_configs/nanollama3.yaml"
 
-optimizer_config: "configs/optimizer_configs/adamw_fused.yaml"
+optimizer_config: "configs/optimizer_configs/adamw.yaml"
 
 eval_config:
   repetition_penalty: 1.0
