+++ tt-train/sources/ttml/metal/optimizers/sgd/device/kernels/compute/sgd_kernel.cpp
@@ -4,9 +4,9 @@
 
 #include <cstdint>
 
-#include "api/compute/compute_kernel_api.h"
 #include "api/compute/bcast.h"
 #include "api/compute/common.h"
+#include "api/compute/compute_kernel_api.h"
 #include "api/compute/eltwise_binary.h"
 #include "tt-train/sources/ttml/metal/common/compute_utils.hpp"
