+++ tt-train/sources/ttml/metal/optimizers/sgd/device/kernels/dataflow/reader_sgd_interleaved_start_id.cpp
