+++ tt-train/sources/ttml/metal/optimizers/sgd/device/sgd_device_operation.hpp
@@ -7,17 +7,17 @@
 #include <optional>
 
 #include "metal/ttnn_all_includes.hpp"
-#include "sgd_fused_device_operation_types.hpp"
-#include "sgd_fused_program_factory.hpp"
+#include "sgd_device_operation_types.hpp"
+#include "sgd_program_factory.hpp"
 
-namespace ttml::metal::optimizers::sgd_fused::device {
+namespace ttml::metal::optimizers::sgd::device {
 
-struct SGDFusedDeviceOperation {
-    using operation_attributes_t = ttml::metal::optimizers::sgd_fused::device::operation_attributes_t;
-    using tensor_args_t = ttml::metal::optimizers::sgd_fused::device::tensor_args_t;
-    using spec_return_value_t = ttml::metal::optimizers::sgd_fused::device::spec_return_value_t;
-    using tensor_return_value_t = ttml::metal::optimizers::sgd_fused::device::tensor_return_value_t;
-    using program_factory_t = std::variant<SGDFusedProgramFactory>;
+struct SGDDeviceOperation {
+    using operation_attributes_t = ttml::metal::optimizers::sgd::device::operation_attributes_t;
+    using tensor_args_t = ttml::metal::optimizers::sgd::device::tensor_args_t;
+    using spec_return_value_t = ttml::metal::optimizers::sgd::device::spec_return_value_t;
+    using tensor_return_value_t = ttml::metal::optimizers::sgd::device::tensor_return_value_t;
+    using program_factory_t = std::variant<SGDProgramFactory>;
 
     static void validate_on_program_cache_miss(const operation_attributes_t&, const tensor_args_t&);
 
@@ -29,11 +29,11 @@ struct SGDFusedDeviceOperation {
     static ttsl::hash::hash_t compute_program_hash(const operation_attributes_t&, const tensor_args_t&);
 };
 
-}  // namespace ttml::metal::optimizers::sgd_fused::device
+}  // namespace ttml::metal::optimizers::sgd::device
 
 namespace ttnn::prim {
 
-ttml::metal::optimizers::sgd_fused::device::SGDFusedDeviceOperation::tensor_return_value_t ttml_sgd_fused(
+ttml::metal::optimizers::sgd::device::SGDDeviceOperation::tensor_return_value_t ttml_sgd(
     const ttnn::Tensor& param,
     const ttnn::Tensor& grad,
     float lr,
