+++ tt-train/sources/ttml/optimizers/adamw.hpp
@@ -1,10 +1,10 @@
-// SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
+// SPDX-FileCopyrightText: (c) 2025 Tenstorrent AI ULC
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#include <core/ttnn_all_includes.hpp>
+#pragma once
 
-#include "optimizer_base.hpp"
+#include "optimizers/optimizer_base.hpp"
 #include "serialization/serializable.hpp"
 
 namespace ttml::optimizers {
@@ -14,48 +14,16 @@ struct AdamWConfig {
     float beta1{0.9F};
     float beta2{0.999F};
     float epsilon{1e-8F};
-    float weight_decay{0.01F};
+    float weight_decay{1e-2F};
     bool amsgrad{false};
-
-    // flag to enable kahan summation to reduce floating point errors
-    bool use_kahan_summation{false};
-};
-
-class MorehAdamW : public OptimizerBase {
-public:
-    [[nodiscard]] std::string get_name() const override {
-        return "MorehAdamW";
-    }
-
-    MorehAdamW(serialization::NamedParameters parameters, const AdamWConfig& config);
-
-    void zero_grad() override;
-
-    void step() override;
-
-    [[nodiscard]] serialization::StateDict get_state_dict() const override;
-    void set_state_dict(const serialization::StateDict& dict) override;
-
-    [[nodiscard]] size_t get_steps() const override;
-    void set_steps(size_t steps) override;
-
-    [[nodiscard]] float get_lr() const override;
-    void set_lr(float lr) override;
-
-private:
-    size_t m_steps{0};
-    AdamWConfig m_config;
-    serialization::NamedParameters m_first_moment;
-    serialization::NamedParameters m_second_moment;
+    bool stochastic_rounding{false};
 };
 
 class AdamW : public OptimizerBase {
 public:
-    [[nodiscard]] std::string get_name() const override {
-        return "AdamW";
-    }
+    [[nodiscard]] std::string get_name() const override;
 
-    AdamW(serialization::NamedParameters parameters, const AdamWConfig& config);
+    explicit AdamW(ttml::serialization::NamedParameters parameters, const AdamWConfig& config);
 
     void zero_grad() override;
 
@@ -68,16 +36,36 @@ public:
     void set_steps(size_t steps) override;
 
     [[nodiscard]] float get_lr() const override;
-
     void set_lr(float lr) override;
 
+    [[nodiscard]] float get_beta1() const;
+    void set_beta1(float beta1);
+
+    [[nodiscard]] float get_beta2() const;
+    void set_beta2(float beta2);
+
+    [[nodiscard]] float get_epsilon() const;
+    void set_epsilon(float epsilon);
+
+    [[nodiscard]] float get_weight_decay() const;
+    void set_weight_decay(float weight_decay);
+
+    [[nodiscard]] bool get_amsgrad() const;
+    void set_amsgrad(bool amsgrad);
+
+    [[nodiscard]] bool get_stochastic_rounding() const;
+    void set_stochastic_rounding(bool stochastic_rounding);
+
 private:
+    void init_max_exp_avg_sq();
+
     size_t m_steps{0};
+    float m_beta1_pow{1.0F};
+    float m_beta2_pow{1.0F};
     AdamWConfig m_config;
-    serialization::NamedParameters m_first_moment;
-    serialization::NamedParameters m_second_moment;
-    serialization::NamedParameters m_max_exp_avg_sq;
-    serialization::NamedParameters m_kahan_compensation;
+    ttml::serialization::NamedParameters m_exp_avg;
+    ttml::serialization::NamedParameters m_exp_avg_sq;
+    ttml::serialization::NamedParameters m_max_exp_avg_sq;
 };
 
 }  // namespace ttml::optimizers
