+++ tt-train/sources/ttml/optimizers/sgd.cpp
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
+// SPDX-FileCopyrightText: (c) 2025 Tenstorrent AI ULC
 //
 // SPDX-License-Identifier: Apache-2.0
 
@@ -10,19 +10,23 @@
 #include "autograd/autocast_tensor.hpp"
 #include "core/debug.hpp"
 #include "core/tt_tensor_utils.hpp"
+#include "metal/operations.hpp"
 #include "serialization/serializable.hpp"
 
 namespace ttml::optimizers {
 
 SGD::SGD(ttml::serialization::NamedParameters parameters, const SGDConfig& config) :
     OptimizerBase(std::move(parameters)), m_config(config) {
-    for (const auto& [name, tensor_ptr] : m_parameters) {
-        if (tensor_ptr->get_requires_grad()) {
-            m_theta.emplace(
-                name,
-                autograd::create_tensor(
-                    core::zeros_like(tensor_ptr->get_value(autograd::PreferredPrecision::FULL)),
-                    /* requires_grad */ false));
+    validate_config();
+    if (m_config.momentum > 0.0) {
+        for (const auto& [name, tensor_ptr] : m_parameters) {
+            if (tensor_ptr->get_requires_grad()) {
+                m_momentum.emplace(
+                    name,
+                    autograd::create_tensor(
+                        core::zeros_like(tensor_ptr->get_value(autograd::PreferredPrecision::FULL)),
+                        /* requires_grad */ false));
+            }
         }
     }
 }
@@ -36,81 +40,55 @@ void SGD::zero_grad() {
 }
 
 void SGD::step() {
+    validate_config();
+
     if (core::debug::Debug::enable_print_tensor_stats()) {
         print_stats();
     }
 
-    for (auto& [name, theta_ptr] : m_theta) {
-        auto theta = theta_ptr->get_value(autograd::PreferredPrecision::FULL);
-        const auto& tensor_ptr = m_parameters.at(name);
-        if (!tensor_ptr->is_grad_initialized()) {
+    for (const auto& [name, theta_ptr] : m_parameters) {
+        if (!theta_ptr->is_grad_initialized()) {
             continue;
         }
+        auto gradients = theta_ptr->get_grad();
+        auto param = theta_ptr->get_value(autograd::PreferredPrecision::FULL);
 
-        auto gradients = tensor_ptr->get_grad();
-
-        if (m_config.weight_decay != 0.0F) {
-            gradients = ttnn::add(
-                ttnn::multiply(
-                    tensor_ptr->get_value(autograd::PreferredPrecision::FULL),
-                    m_config.weight_decay,
-                    /* fast_and_approximate_mode*/ true),
-                gradients);
+        std::optional<ttnn::Tensor> momentum_buffer;
+        if (m_config.momentum > 0.0) {
+            auto it = m_momentum.find(name);
+            if (it == m_momentum.end()) {
+                auto buf = autograd::create_tensor(
+                    core::zeros_like(param),
+                    /* requires_grad */ false);
+                it = m_momentum.emplace(name, std::move(buf)).first;
+            }
+            momentum_buffer = it->second->get_value(autograd::PreferredPrecision::FULL);
         }
 
-        if (m_config.momentum != 0.0F) {
-            if (m_steps != 0) {
-                // apply momentum
-                theta = ttnn::multiply(
-                    theta,
-                    m_config.momentum,
-                    /* fast_and_approximate_mode*/ true);
-                // dampening
-                if (m_config.dampening != 0.0F) {
-                    theta = ttnn::add(
-                        theta,
-                        ttnn::multiply(
-                            gradients,
-                            1 - m_config.dampening,
-                            /* fast_and_approximate_mode*/ true));
-                } else {
-                    theta = ttnn::add(theta, gradients);
-                }
-            } else {
-                theta = ttnn::add(theta, gradients);
-            }
-
-            if (m_config.nesterov) {
-                gradients = ttnn::add(
-                    gradients,
-                    ttnn::multiply(
-                        theta,
-                        m_config.momentum,
-                        /* fast_and_approximate_mode*/ true));
-            } else {
-                gradients = theta;
-            }
-        }
-        theta_ptr->set_value(theta);
-        tensor_ptr->set_value(ttnn::subtract(
-            tensor_ptr->get_value(autograd::PreferredPrecision::FULL),
-            ttnn::multiply(
-                gradients,
-                m_config.lr,
-                /* fast_and_approximate_mode*/ true)));
+        // The first step should not apply dampening
+        float dampening = m_steps == 0 ? 0.0f : m_config.dampening;
+        ttml::metal::sgd(
+            param,
+            gradients,
+            m_config.lr,
+            m_config.momentum,
+            dampening,
+            m_config.weight_decay,
+            m_config.nesterov,
+            momentum_buffer);
     }
     m_steps++;
 }
 
 serialization::StateDict SGD::get_state_dict() const {
     serialization::StateDict dict;
-    dict["theta"] = m_theta;
+    dict["momentum"] = m_momentum;
     dict["steps"] = m_steps;
     return dict;
 }
 
 void SGD::set_state_dict(const serialization::StateDict& dict) {
-    m_theta = std::get<serialization::NamedParameters>(dict.at("theta"));
+    m_momentum = std::get<serialization::NamedParameters>(dict.at("momentum"));
     m_steps = serialization::get_value_type<size_t>(dict, "steps");
 }
 
@@ -120,6 +98,62 @@ size_t SGD::get_steps() const {
 
 void SGD::set_steps(size_t steps) {
     this->m_steps = steps;
+}
+
+float SGD::get_lr() const {
+    return m_config.lr;
+}
+
+void SGD::set_lr(float lr) {
+    m_config.lr = lr;
+}
+
+float SGD::get_momentum() const {
+    return m_config.momentum;
+}
+
+void SGD::set_momentum(float momentum) {
+    m_config.momentum = momentum;
+}
+
+float SGD::get_dampening() const {
+    return m_config.dampening;
+}
+
+void SGD::set_dampening(float dampening) {
+    m_config.dampening = dampening;
+}
+
+float SGD::get_weight_decay() const {
+    return m_config.weight_decay;
+}
+
+void SGD::set_weight_decay(float weight_decay) {
+    m_config.weight_decay = weight_decay;
+}
+
+bool SGD::get_nesterov() const {
+    return m_config.nesterov;
+}
+
+void SGD::set_nesterov(bool nesterov) {
+    m_config.nesterov = nesterov;
+}
+
+void SGD::validate_config() const {
+    if (m_config.nesterov) {
+        if (m_config.dampening != 0.0) {
+            throw std::runtime_error(
+                fmt::format("Nesterov momentum requires zero dampening! dampening={}", m_config.dampening));
+        }
+        if (m_config.momentum <= 0.0) {
+            throw std::runtime_error(
+                fmt::format("Nesterov momentum requires a positive momentum! momentum={}", m_config.momentum));
+        }
+    }
+    if (m_config.dampening != 0.0 && m_config.momentum == 0.0) {
+        throw std::runtime_error(fmt::format("Dampening requires a positive momentum! momentum={}", m_config.momentum));
+    }
 }
 
 }  // namespace ttml::optimizers
