+++ tt-train/sources/ttml/optimizers/sgd.hpp
@@ -1,4 +1,4 @@
-// SPDX-FileCopyrightText: © 2024 Tenstorrent AI ULC
+// SPDX-FileCopyrightText: (c) 2025 Tenstorrent AI ULC
 //
 // SPDX-License-Identifier: Apache-2.0
 
@@ -35,18 +35,27 @@ public:
     [[nodiscard]] size_t get_steps() const override;
     void set_steps(size_t steps) override;
 
-    [[nodiscard]] float get_lr() const override {
-        return m_config.lr;
-    }
+    [[nodiscard]] float get_lr() const override;
 
-    void set_lr(float lr) override {
-        m_config.lr = lr;
-    }
+    void set_lr(float lr) override;
+
+    [[nodiscard]] float get_momentum() const;
+    void set_momentum(float momentum);
+
+    [[nodiscard]] float get_dampening() const;
+    void set_dampening(float dampening);
+
+    [[nodiscard]] float get_weight_decay() const;
+    void set_weight_decay(float weight_decay);
+
+    [[nodiscard]] bool get_nesterov() const;
+    void set_nesterov(bool nesterov);
 
 private:
+    void validate_config() const;
     size_t m_steps{0};
     SGDConfig m_config;
-    ttml::serialization::NamedParameters m_theta;
+    ttml::serialization::NamedParameters m_momentum;
 };
 
 }  // namespace ttml::optimizers
