+++ tt-train/tests/CMakeLists.txt
@@ -32,9 +32,9 @@ set(SOURCES
     datasets/utils_test.cpp
     datasets/generators_test.cpp
     datasets/dataloader_test.cpp
+    optimizers/adamw_composite_test.cpp
+    optimizers/sgd_test.cpp
     optimizers/adamw_test.cpp
-    optimizers/sgd_fused_test.cpp
-    optimizers/adamw_fused_test.cpp
     optimizers/adamw_full_precision_test.cpp
     optimizers/no_op_test.cpp
     modules/distributed/linear_test.cpp
